move tracing executor to its own file
